feat(model): support default model selection

+++ internal/application/service/model_delete_test.go
@@ -100,7 +100,7 @@ func (s *stubModelRepoForDelete) Delete(_ context.Context, _ uint64, id string) 
 	}
 	return nil
 }
-func (s *stubModelRepoForDelete) ClearDefaultByType(context.Context, uint, types.ModelType, string) error {
+func (s *stubModelRepoForDelete) ClearDefaultByType(context.Context, uint64, types.ModelType, string) error {
 	return nil
 }
